Fixed integer sign mismatch.

+++ Core/Source/DTTextHTMLElement.m
@@ -23,7 +23,7 @@
 - (void)_appendHTMLToString:(NSMutableString *)string indentLevel:(NSUInteger)indentLevel
 {
 	// indent to the level
-	for (int i=0; i<indentLevel; i++)
+	for (NSUInteger i=0; i<indentLevel; i++)
 	{
 		[string appendString:@"   "];
 	}
